Changed tilemap rendering.
The tilemap now automatically updates each layer when they are edited at the end of the frame.

This is also an optimization, as first I'd upload the entire tilemap to opengl, now I just upload each layer, which might be slower actually if we updated more layers, might want to profile this behavior later.